requirements.txt
black
flask
fastapi
google-cloud-aiplatform
google-cloud-storage
uvicorn
Pillow